Remove the need for commons-configuration.
