+++ kie-wb-common-ala/kie-wb-common-ala-spi/src/main/java/org/guvnor/ala/pipeline/execution/PipelineExecutorError.java
@@ -16,7 +16,7 @@
 
 package org.guvnor.ala.pipeline.execution;
 
-import org.apache.commons.lang.exception.ExceptionUtils;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 
 public class PipelineExecutorError {
