+++ kie-wb-common-ala/kie-wb-common-ala-spi/src/main/java/org/guvnor/ala/util/VariableInterpolation.java
@@ -15,7 +15,6 @@
  */
 package org.guvnor.ala.util;
 
-import java.beans.PropertyDescriptor;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.util.Map;
@@ -24,10 +23,7 @@ import net.bytebuddy.ByteBuddy;
 import net.bytebuddy.dynamic.loading.ClassLoadingStrategy;
 import net.bytebuddy.implementation.InvocationHandlerAdapter;
 import net.bytebuddy.matcher.ElementMatchers;
-import org.apache.commons.beanutils.PropertyUtilsBean;
-import org.apache.commons.configuration.interpol.ConfigurationInterpolator;
-import org.apache.commons.lang.text.StrLookup;
-import org.apache.commons.lang.text.StrSubstitutor;
+import org.apache.commons.lang3.text.StrSubstitutor;
 import org.guvnor.ala.config.CloneableConfig;
 
 /**
@@ -41,49 +37,11 @@ public final class VariableInterpolation {
 
     }
 
-    private static final ConfigurationInterpolator interpolator = new ConfigurationInterpolator();
-    private static final StrSubstitutor substitutor = new StrSubstitutor(interpolator);
+    private static final StrSubstitutor substitutor = new StrSubstitutor();
 
-    public static <T> T interpolate(final Map<String, Object> values,
-                                    final T object) {
-        interpolator.setDefaultLookup(new MapOfMapStrLookup(values));
+    public static <T> T interpolate(final Map<String, Object> values, final T object) {
+        substitutor.setVariableResolver(new VariableLookup(values));
         return proxy(object);
-    }
-
-    private static class MapOfMapStrLookup extends StrLookup {
-
-        private final Map map;
-
-        MapOfMapStrLookup(Map map) {
-            this.map = map;
-        }
-
-        @Override
-        public String lookup(String key) {
-            if (this.map == null) {
-                return null;
-            } else {
-                int dotIndex = key.indexOf(".");
-                Object obj = this.map.get(key.substring(0,
-                                                        dotIndex < 0 ? key.length() : dotIndex));
-                if (obj instanceof Map) {
-                    return new MapOfMapStrLookup(((Map) obj)).lookup(key.substring(key.indexOf(".") + 1));
-                } else if (obj != null && !(obj instanceof String) && key.contains(".")) {
-                    final String subkey = key.substring(key.indexOf(".") + 1);
-                    for (PropertyDescriptor descriptor : new PropertyUtilsBean().getPropertyDescriptors(obj)) {
-                        if (descriptor.getName().equals(subkey)) {
-                            try {
-                                return descriptor.getReadMethod().invoke(obj).toString();
-                            } catch (Exception ex) {
-                                continue;
-                            }
-                        }
-                    }
-                }
-
-                return obj == null ? "" : obj.toString();
-            }
-        }
     }
 
     public static <T> T proxy(final T instance) {
